+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.10] - 2017-10-05
 ### Added
 - Additional actions for RSVPs, including `getEventRsvp`, `listEventRsvps`, and `updateEventRsvp`.
 
@@ -79,7 +81,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 - Initial checkin.
 
-[Unreleased]: https://github.com/CultureHQ/client/compare/v0.0.8...HEAD
+[Unreleased]: https://github.com/CultureHQ/client/compare/v0.0.10...HEAD
+[0.0.8]: https://github.com/CultureHQ/client/compare/v0.0.9...v0.0.10
+[0.0.8]: https://github.com/CultureHQ/client/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/CultureHQ/client/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/CultureHQ/client/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/CultureHQ/client/compare/v0.0.5...v0.0.6